requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.4
requests==2.32.0
python-dotenv==1.0.0
pandas

